Modified lookup_value to also use lookup_url_kwarg
The RetrieveSqlQueryKeyBit use to fetch the lookup_value only looking at lookup_field variable of the view. DRF allows since 8 years to supply for views a separate lookup_url_kwarg property, for when the name of the URL kwarg is different from the lookup_field name. DRF will use lookup_field to get the lookup_value from the URL unless lookup_url_kwarg is defined. This commit adds the same logic to the RetrieveSqlQueryKeyBit, allowing this bit to work when lookup_url_kwarg is used in DRF.